\docType{data}
\name{exampleCounts}
\alias{exampleCounts}
\title{An example data generated by MultiRankSeq package}
\description{
  This data set is a random counts data based on negative
  binomial distribution. The codes for generating this data
  can be found on the example of findDiff function.
}
\keyword{datasets}

