\docType{data}
\name{exampleCountsDiff}
\alias{exampleCountsDiff}
\title{An example data generated by MultiRankSeq package}
\description{
  This data set is the results from DESeq, edgeR, baySeq
  packages, which is based on the exampleCounts data set.
  The codes for generating this data can be found on the
  example of findDiff function.
}
\keyword{datasets}

