% Generated by roxygen2 (4.0.2): do not edit by hand
\name{MultiRankSeqReport}
\alias{MultiRankSeqReport}
\title{MultiRankSeqReport}
\usage{
MultiRankSeqReport(output = "MultiRankSeq.html", rawCounts, group = c(0, 0,
  0, 1, 1, 1), diffResult, seed = NULL, paired = NULL, exportCounts = F,
  useVStCount = T, useAllInHeatmap = F, percent = 0.05, ...)
}
\arguments{
\item{output}{the path and name for report file.}

\item{rawCounts}{the count data used in analysis, which should be a matrix or a data.frame.}

\item{group}{the group of count data, which should be 0 or 1. The samples labeled as 1 will be compared with samples labeled as 0.}

\item{diffResult}{The result generated by \code{\link{findDiff}} function. If not specified, findDiff function will be used to generate it.}

\item{seed}{the random seed for bayseq. The user can set it so that the bayseq result can be reproduced.}

\item{paired}{the pairs of count data. Will be NULL if no pair. For example, if sample 1 and 2 are one pair, sample 3 and 4 are another pair, the paired should be c(1,1,2,2).}

\item{exportCounts}{Logical. Indicate if the raw counts will be exported in result table.}

\item{useVStCount}{Logical. The DESeq2 package can apply a variance stabilizing transformation (VST) to count data, which can be used in checking outliers or clustering. This parameter indicates if the transformed counts will be used in heatmap, boxplot, and correlation in reports.}

\item{useAllInHeatmap}{Logical.}

\item{percent}{The percent of genes with highest SD, to be used in heatmap.}

\item{...}{Other parameters used in \code{\link{diffReport}} function.}
}
\description{
A wrapper for function findDiff and diffReport.
}
\details{
A function to generate MultiRankSeq report based on counts data.
}
\examples{
#A simple example
data(exampleCounts)
data(exampleCountsDiff)
result<-MultiRankSeqReport(output="MultiRankSeq1.html",diffResult=exampleCountsDiff,rawCounts=exampleCounts)
\dontrun{
#An example: from counts data to report, paired data
#this code may take two minutes, because performing baySeq, DESeq2, and edgeR may be slow.
result<-MultiRankSeqReport(output="MultiRankSeq2.html",rawCounts=exampleCounts,group=c(0,0,0,1,1,1),paired=c(1:3,1:3)
}
}

