\name{diffReport}
\alias{diffReport}
\title{diffReport}
\usage{
diffReport(input = system.file("extdata", "report.Rhtml", package =
  "MultiRankSeq"), output, diffResult, rawCounts, group = c(0, 0, 0, 1, 1, 1))
}
\arguments{
  \item{input}{The templet file. You don't need to specify
  it in most cases. A file in MultiRankSeq package will be
  used.}

  \item{diffResult}{The result generated by
  \code{\link{findDiff}} function.}

  \item{output}{the path and name for report file.}

  \item{rawCounts}{the count data used in analysis, which
  should be a matrix or a data.frame.}

  \item{group}{the group of count data, which should be 0
  or 1. The samples labeled as 1 will be compared with
  samples labeled as 0.}
}
\description{
The function will generate a report based on the result of
DESeq, edgeR, and baySeq analysis.
}
\examples{
data(exampleCounts)
data(exampleCountsDiff)
diffReport(output="MultiRankSeq.html",diffResult=exampleCountsDiff,rawCounts=exampleCounts)
}

