% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiRankSeqFunctions.R
\docType{data}
\name{exampleCounts}
\alias{exampleCounts}
\title{An example data generated by MultiRankSeq package}
\usage{
exampleCounts
}
\description{
This data set is a random counts data based on negative binomial distribution. The codes for generating this data can be found on the example of findDiff function.
}

