% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiRankSeqFunctions.R
\docType{data}
\name{exampleCountsDiff}
\alias{exampleCountsDiff}
\title{An example data generated by MultiRankSeq package}
\usage{
exampleCountsDiff
}
\description{
This data set is the results from DESeq, edgeR, baySeq packages, which is based on the exampleCounts data set. The codes for generating this data can be found on the example of findDiff function.
}

