\name{MultiRankSeqReport}
\alias{MultiRankSeqReport}
\title{MultiRankSeqReport}
\usage{
MultiRankSeqReport(output = "MultiRankSeq.html", rawCounts, group = c(0, 0,
  0, 1, 1, 1), diffResult, ...)
}
\arguments{
  \item{output}{the path and name for report file.}

  \item{rawCounts}{the count data used in analysis, which
  should be a matrix or a data.frame.}

  \item{group}{the group of count data, which should be 0
  or 1. The samples labeled as 1 will be compared with
  samples labeled as 0.}

  \item{diffResult}{The result generated by
  \code{\link{findDiff}} function. If not specified,
  findDiff function will be used to generate it.}

  \item{...}{Other parameters used in
  \code{\link{diffReport}} function.}
}
\description{
A wrapper for function findDiff and diffReport.
}
\details{
A function to generate MultiRankSeq report based on counts
data.
}
\examples{
#A simple example
data(exampleCounts)
data(exampleCountsDiff)
MultiRankSeqReport(output="MultiRankSeq1.html",diffResult=exampleCountsDiff,rawCounts=exampleCounts)
\dontrun{
#An example: from counts data to report.
#this code may take two minutes, because performing baySeq, DESeq, and edgeR may be slow.
MultiRankSeqReport(output="MultiRankSeq2.html",rawCounts=exampleCounts)
}
}

